+++ src/signatures.rs
@@ -423,6 +423,13 @@ signatures! {
     value = b"S      1\n" offset = 72
     value = b"S0000001\n" offset = 72
 
+    format = MicrosoftWorksDatabase
+    value = b"\x20\x54\x02\x00\x00\x00\x05\x54\x02\x00"
+
+    format = MicrosoftWorksSpreadsheet
+    value = b"\x00\x00\x02\x00\x04\x04\x05\x54\x02\x00"
+    value = b"\xFF\x00\x02\x00\x04\x04\x05\x54\x02\x00"
+
     format = OggMedia
     value = b"OggS", b"\x01\x76\x69\x64\x65\x6F" offset = 28
 
@@ -758,6 +765,10 @@ signatures! {
     format = Mp3Url
     value = b"#EXTM3U"
 
+    format = MpegDashManifest
+    value = b"\xEF\xBB\xBF<MPD"
+    value = b"<MPD"
+
     format = ReallySimpleSyndication
     value = b"\xEF\xBB\xBF<rss"
     value = b"<rss"
@@ -1065,6 +1076,9 @@ signatures! {
 
     format = MicrosoftDirectdrawSurface
     value = b"DDS "
+
+    format = MicrosoftWorksWordProcessor
+    value = b"\x01\xFE", b"\x01\x00" offset = 112
 
     format = Model3dBinary
     value = b"3DMO"